requirements.txt
Flask==2.1.1
PyMuPDF==1.18.15
python-docx==0.8.11
spacy==3.2.0
nltk==3.6.5
scikit-learn==1.0.2
pandas==1.3.3
scikit-learn==1.0.2
wordcloud==1.8.1
openpyxl==3.0.9

